+++ CHANGELOG.md
@@ -44,7 +44,8 @@
 - Additional simulation variables: flap positions, water ruder and brakes
 - Skip to begin, end, forward and backward buttons
 - Settings dialog with sample rate settings
-- Basic CSV export
+- Basic CSV import & export
+- Window stay on top option
 
 ### Bug Fixes